+++ R/2-papply.R
@@ -6,7 +6,7 @@
 #'  For i,j in 1...N rows (or columns), it returns an N by N matrix
 #'  containing the result of FUN(Xi, Xj).
 #'
-#' @param X A matrix.
+#' @param X Any matrix or data.frame.
 #' @param MARGIN Choose from 1 or 2. If 1, apply function to
 #'  pairs of rows. If 2, apply function to pairs of columns.
 #' @param FUN The function to apply to row or column pairs.
@@ -14,6 +14,7 @@
 #' @param lower_only A boolean. Toggles whether to apply
 #'  function to lower triangle pairs only.
 #' @param ... Arguments passed to \code{FUN}.
+#'
 #' @return A square matrix.
 #'
 #' @examples
@@ -28,25 +29,20 @@ papply <- function(X, MARGIN = 2, FUN, lower_only = FALSE, ...){
     X <- t(X)
   }
 
-  args <- as.list(substitute(list(...)))[-1]
-
   # Apply difference function
   D = matrix(0, ncol(X), ncol(X))
   for(i in 1:ncol(X)){
     numTicks <- progress(i, ncol(X), numTicks)
     for(j in 1:ncol(X)){
       if(lower_only){
-        if(j<i){
-          D[i,j] <- do.call(FUN, append(list(X[,i], X[,j]), args))
-        }
+        if(j<i) D[i,j] <- do_call(FUN, X[,i], X[,j], ...)
       }else{
-        D[i,j] <- do.call(FUN, append(list(X[,i], X[,j]), args))
+        D[i,j] <- do_call(FUN, X[,i], X[,j], ...)
       }
     }
   }
 
-  # Name results
   rownames(D) <- colnames(X)
   colnames(D) <- colnames(X)
-  return(D)
+  D
 }